+++ src/main/java/com/appsalud/plataformaSalud/controladores/PacienteControlador.java
@@ -1,10 +1,33 @@
 package com.appsalud.plataformaSalud.controladores;
 
+import com.appsalud.plataformaSalud.enumeraciones.Especialidad;
+import com.appsalud.plataformaSalud.enumeraciones.ObraSocial;
+import com.appsalud.plataformaSalud.servicios.UsuarioPacienteServicio;
+import com.appsalud.plataformaSalud.servicios.UsuarioProfesionalServicio;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping("/paciente")
 public class PacienteControlador {
-    
+    @Autowired
+    private UsuarioPacienteServicio usuarioPacienteServicio;
+    @GetMapping("/home")
+    public String profesionalHome() {
+        return "paciente.html";
+    }
+
+    @GetMapping("/registrarPaciente")
+    public String registroProfesional(Model model) {
+        List<ObraSocial> listaObrasSociales = Arrays.stream(ObraSocial.values()).collect(Collectors.toList());
+        model.addAttribute("listaObrasSociales", listaObrasSociales);
+        return "registroProfesional.html";
+    }
 }
