+++ src/main/java/com/appsalud/plataformaSalud/repositorios/UsuarioRepositorio.java
@@ -1,6 +1,7 @@
 package com.appsalud.plataformaSalud.repositorios;
 
 import com.appsalud.plataformaSalud.entidades.Usuario;
+import com.appsalud.plataformaSalud.entidades.UsuarioPaciente;
 import com.appsalud.plataformaSalud.entidades.UsuarioProfesional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -16,4 +17,6 @@ public interface UsuarioRepositorio extends JpaRepository<Usuario, String>{
 
     @Query("SELECT p FROM Usuario p WHERE p.id = :id")
     public Usuario buscarPorId(@Param("id") String id);
+    @Query("SELECT u FROM Usuario u WHERE u.email = :email")
+    public Optional<UsuarioPaciente> buscarPorEmailPaciente(@Param("email") String email);
 }
